requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
fake-useragent==0.1.11
fastapi==0.59.0
h11==0.9.0
httptools==0.1.1
idna==2.10
lxml==4.5.2
parse==1.15.0
pydantic==1.6.1
pyee==7.0.2
pyppeteer==0.2.2
pyquery==1.4.1
requests==2.24.0
requests-html==0.10.0
six==1.15.0
soupsieve==2.0.1
starlette==0.13.4
tqdm==4.48.0
urllib3==1.25.9
uvicorn==0.11.7
uvloop==0.14.0
w3lib==1.22.0
websockets==8.1
tinydb-4.1.1
